Add helper asserts
